+++ autoload.php
@@ -18,35 +18,3 @@ spl_autoload_register(function ($class_name) {
         require_once $file_path;
     }
 });
-
-/*
-//OLD STYLE
-require_once __DIR__ . '/core/LibConsoleColor.php';
-require_once __DIR__ . '/core/LibLog.php';
-require_once __DIR__.'/core/Spirit.php';
-require_once __DIR__.'/core/Enoch.php';
-require_once __DIR__.'/core/Walker.php';
-require_once __DIR__.'/core/LibMySQL.php';
-require_once __DIR__.'/core/LibSFTP.php';
-require_once __DIR__ . '/core/LibSession.php';
-
-require_once __DIR__.'/SmallPHPMail/phpmailerException.php';
-require_once __DIR__.'/SmallPHPMail/PHPMailer.php';
-require_once __DIR__.'/SmallPHPMail/SMTP.php';
-
-require_once __DIR__.'/core/LibMail.php';
-
-require_once __DIR__ . '/mvc/BaseCodedException.php';
-require_once __DIR__ . '/mvc/ApiInterface.php';
-require_once __DIR__ . '/mvc/Lamech.php';
-require_once __DIR__ . '/mvc/RouterInterface.php';
-require_once __DIR__ . '/mvc/Naamah.php';
-require_once __DIR__ . '/mvc/Adah.php';
-require_once __DIR__ . '/mvc/MiddlewareInterface.php';
-
-require_once __DIR__ . '/service/QueueItem.php';
-require_once __DIR__ . '/service/QueueInterface.php';
-
-require_once __DIR__ . '/service/CacheInterface.php';
-require_once __DIR__ . '/service/FileCache.php';
-*/