+++ test/WalkingSystem/walk.php
@@ -5,9 +5,8 @@
  * Date: 2017/3/27
  * Time: 10:26
  */
-
+require_once __DIR__ . '/../../vendor/autoload.php';
 require_once __DIR__ . '/../../autoload.php';
-
 require_once __DIR__ . '/sample/SampleEnoch.php';
 
 date_default_timezone_set("Asia/Shanghai");